pyproject.toml
[project]
name = "agentsentry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"pydantic>=2.7.0",
	"openai>=1.40.0",
	"fastapi>=0.111.0",
	"uvicorn>=0.30.0",
	"python-dotenv>=1.0.1",
	"mysql-connector-python>=9.0.0",
]
